- Refactored Github mock response generation logic into   GithubApiTestMixin - Added support for ending a parser section when a new top level heading   is encountered
